Refactor: 커뮤니티 관련 fragment ui/view 폴더로 이동 #55
