+++ app/src/main/java/com/project/veganlife/community/ui/view/CommunityHomeFragment.kt
@@ -1,4 +1,4 @@
-package com.project.veganlife.community
+package com.project.veganlife.community.ui.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
