+++ app/src/main/res/layout/fragment_community_search.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/sub_gray1"
-    tools:context=".community.CommunitySearchFragment">
+    tools:context=".community.ui.view.CommunitySearchFragment">
 
     <include
         android:id="@+id/include_community_search_toolbar"
