+++ app/src/main/res/layout/fragment_community_write_edit_feed.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/gray1"
-    tools:context=".community.CommunityWriteFeedFragment">
+    tools:context=".community.ui.view.CommunityWriteFeedFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar_community_write_edit_feed"
